// app/models/UserActivitySettings.ts
class UserActivitySettings {
    name: string;
    isCurrent: boolean;
    flashcards: any;
    whiteboard: any;
    timedReading: any;
    wordcards: any;
    lettercards: any;
    blackboard: any;
    blankTiles: any;
    workbookTiles: any;
    miniTiles: any;
    spelling: any;
    sentences: any;
    phrases: any;
    passages: any;
    pdfViewer: any;
    ownerKey: string;
    _id: string;
}

export {UserActivitySettings}